requirements.txt
aiogram==3.22.0
tortoise-orm==0.25.1
aiohttp==3.12.15
python-decouple==3.8
parsel==1.10.0
asyncpg==0.30.0
betterlogging==1.0.0
openpyxl==3.1.5
apscheduler==3.11.0
pandas==2.3.2
playwright==1.58.0
aerich==0.7.2